Refactored the Add Property and Edit Property forms to be server components using server actions. Didn't delete the original API routes yet. Moved the form components into appropriate directories. Tweaked some comments, got rid of TODOs, made the Amenities checkboxes a component and put the list of amenities into a JS object like I did for US States.
