+++ app/actions/editProperty.js
@@ -1,33 +1,36 @@
 'use server';
-import connectDB from "@/config/database";
+import connectDB from '@/config/database';
 import Property from '@/models/Property';
-import { getSessionUser } from "@/utils/getSessionUser";
-import cloudinary from "@/config/cloudinary";
-import { revalidatePath } from "next/cache";
+import { getSessionUser } from '@/utils/getSessionUser';
+import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 
-async function addProperty( formData ) {
+async function editProperty( propertyID, formData ) {
   await connectDB();
 
   //#region User Authentication & Authorization
   const sessionUser = await getSessionUser();
 
-  // NOTE: throwing an Error from our server actions will be caught by our
-  // error.jsx ErrorBoundary component and show the user an Error page with
-  // message of the thrown error.
-
   if ( ! sessionUser || ! sessionUser.userID ) {
-    throw new Error( 'User ID is required' );
+    return new Response( 'User ID is required.', { status: 401 } );
   }
 
   const { userID } = sessionUser;
   //#endregion
 
-  //#region Create Property data object for database
-  // Get Amenities & Image Data
-  const amenities = formData.getAll( 'amenities' );
-  const images = formData.getAll( 'images' ).filter( ( image ) => image.name !== '' );
+  // Get existing property by ID
+  const existingProperty = await Property.findById( propertyID );
 
+  // Verify ownership
+  if ( existingProperty.owner.toString() !== userID ) {
+    return new Response("Unauthorized: user doesn't own requested property.", { status: 401 });
+  }
+
+  //#region Create Property data object for database
+  // Get Amenities
+  const amenities = formData.getAll( 'amenities' );
+
+  // Create big ol' object to ship to DB.
   // NOTE: Images aren't set directly here since they're grabbed from the form data,
   // uploaded to Cloudinary, then the URLs for them from Cloudinary are added to the
   // images array in the DB for the property.
@@ -55,49 +58,21 @@ async function addProperty( formData ) {
       email: formData.get( 'seller_info.email' ),
       phone: formData.get( 'seller_info.phone' ),
     },
-    images: [],
     owner: userID,
   };
-
-  //#region Upload images to Cloudinary
-  for ( const image of images ) {
-    const mimeType = image.type;
-    const imageBuffer = await image.arrayBuffer();
-    const imageArray = Array.from( new Uint8Array( imageBuffer ) );
-    const imageData = Buffer.from( imageArray );
-
-    // Convert image data to base64
-    const imageBase64 = imageData.toString( 'base64' );
-
-    // Make Cloudinary upload request
-    const result = await cloudinary.uploader.upload(
-      `data:${ mimeType };base64,${ imageBase64 }`,
-      {
-        folder: 'propertypulse',
-      }
-    );
-
-    // Add each uploaded image's secure URL to the propertyData.images array
-    propertyData.images.push( result.secure_url );
-  }
-  //#endregion
   //#endregion
 
-  //#region Create new Property from schema and save to DB
-  const newProperty = new Property( propertyData );
+  //#region Update Property in DB with new data
+  const updatedProperty = await Property.findByIdAndUpdate(propertyID, propertyData);
 
-  await newProperty.save();
-  //#endregion
-
-  //#region Cache Invalidation & Server Response
   // NOTE: Invalidate cache so that the image is fetched for the new property when
   // user next visits /properties. Since properties are on most pages, we can just
   // revalidate everything using the top level layout.
   revalidatePath( '/', 'layout' );
 
   // Redirect to new single Property
-  redirect( `/properties/${ newProperty._id }` );
+  redirect( `/properties/${ updatedProperty._id }` );
   //#endregion
 }
 
-export default addProperty;
+export default editProperty;
