+++ utils/amenitiesList.js
@@ -1,0 +1,24 @@
+const allAmenities = {
+  'Air Conditioning': 'air_conditioning',
+  'Wifi': 'wifi',
+  'Kitchinette': 'kitchinette',
+  'Full Kitchen': 'kitchen',
+  'Dishwasher': 'dishwasher',
+  'Coffee Maker': 'coffee_maker',
+  'BBQ/Grill': 'bbq_grill',
+  'Washer & Dryer': 'washer_dryer',
+  'Smart TV': 'smart_tv',
+  'Pet Friendly': 'pet_friendly',
+  'Free Parking': 'free_parking',
+  'Swimming Pool': 'pool',
+  'Hot Tub': 'hot_tub',
+  'Fireplace': 'fireplace',
+  '24/7 Security': '24_7_security',
+  'Wheelchair Accessible': 'wheelchair_accessible',
+  'Elevator Access': 'elevator_access',
+  'Gym/Fitness Center': 'gym_fitness_center',
+  'Balcony/Patio': 'balcony_patio',
+  'EV Charging': 'ev_charging',
+};
+
+export default allAmenities;
